moved federal_agencies and scraper in abstract scraper directory
