+++ abstract/federal_agencies/federal_scraper.py
@@ -1,13 +1,50 @@
-from federal_utils import (
-    scrape_federal_agency,
+from abstract.utils import (
     DEFAULT_FIELDS,
     send_doc_to_kafka,
     init_kafka_producer,
+    get_all_documents_recurse,
 )
 import os
 import sys
 import json
 import argparse
+
+
+def scrape_federal_agency(
+    fields: dict = DEFAULT_FIELDS,
+    exact_date: str = None,
+    start_date: str = None,
+    extra_params: dict = None,
+):
+    """
+    Fetch executive orders from the Federal Register API. By default, scrapes all available orders from 1994 onwards.
+    exact_date or start_date are optional, though mutually exclusive. If both are provided, only exact_date will be used.
+
+    :param fields: List of fields to include in the response.
+    :param exact_date: Fetch orders from a specific date (YYYY-MM-DD).
+    :param start_date: Fetch orders from this date onwards (YYYY-MM-DD).
+    :return: List of executive orders.
+    """
+    base_url = "https://www.federalregister.gov/api/v1/documents.json"
+
+    params = {
+        "order": "oldest",
+        "per_page": 1000,
+    }
+
+    if exact_date:
+        params["conditions[publication_date][is]"] = exact_date
+    elif start_date:  # elif since start_date and exact_date are mutually exclusive
+        params["conditions[publication_date][gte]"] = start_date
+
+    if extra_params:
+        for key, value in extra_params.items():
+            params[key] = value
+
+    if fields:
+        params["fields[]"] = fields
+
+    return get_all_documents_recurse(base_url, params)
 
 
 def write_out_scrape(
@@ -36,7 +73,9 @@ def write_out_scrape(
             )
     else:
         for file in scraped_jsons:
-            document_title_value = file.get(document_title, file.get("document_number", "unknown"))
+            document_title_value = file.get(
+                document_title, file.get("document_number", "unknown")
+            )
             with open(f"{output_dir}/{document_title_value}.json", "w") as f:
                 f.write(json.dumps(file, indent=2))
         print(f"Wrote {len(scraped_jsons)} orders to individual files", file=sys.stderr)
@@ -84,7 +123,7 @@ def main():
     parser.add_argument(
         "--kafka",
         help="If set, will send the documents to Kafka at the given topic instead of writing them out (i.e '--kafka <topic>')",
-        default=None
+        default=None,
     )
 
     parser.add_argument(
@@ -111,7 +150,7 @@ def main():
             send_doc_to_kafka(doc, topic=args.kafka, kakfa_producer=kakfa_producer)
             print(f"Sent document {doc.get(args.document_title)} to Kafka")
         return
-    
+
     else:
         write_out_scrape(
             scraped_documents, args.output_dir, args.group, args.document_title
